Split routes into dedicated routers
