+++ app/database.py
@@ -1,6 +1,8 @@
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy.orm import sessionmaker, Session
+
+from typing import Generator
 
 # SQLite DB URL
 SQLALCHEMY_DATABASE_URL = "sqlite:///./test.db"
@@ -8,3 +10,12 @@ SQLALCHEMY_DATABASE_URL = "sqlite:///./test.db"
 engine = create_engine(SQLALCHEMY_DATABASE_URL, connect_args={"check_same_thread": False})
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 Base = declarative_base()
+
+
+def get_db() -> Generator[Session, None, None]:
+    """Yield a database session and ensure it is closed."""
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
