+++ app/routers/posts.py
@@ -1,0 +1,62 @@
+from typing import List
+
+from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.orm import Session
+
+from .. import auth, models, schemas
+from ..database import get_db
+
+router = APIRouter(prefix="/posts", tags=["posts"])
+
+@router.post("/", response_model=schemas.Post)
+def create_post(
+    post: schemas.PostCreate,
+    current_user: models.User = Depends(auth.get_current_user),
+    db: Session = Depends(get_db),
+):
+    db_post = models.Post(**post.dict(), owner_id=current_user.id)
+    db.add(db_post)
+    db.commit()
+    db.refresh(db_post)
+    return db_post
+
+
+@router.get("/", response_model=List[schemas.Post])
+def read_posts(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
+    posts = db.query(models.Post).offset(skip).limit(limit).all()
+    return posts
+
+
+@router.get("/my/", response_model=List[schemas.Post])
+def read_my_posts(
+    current_user: models.User = Depends(auth.get_current_user),
+    db: Session = Depends(get_db),
+):
+    posts = db.query(models.Post).filter(models.Post.owner_id == current_user.id).all()
+    return posts
+
+
+@router.get("/{post_id}", response_model=schemas.Post)
+def read_post(post_id: int, db: Session = Depends(get_db)):
+    post = db.query(models.Post).filter(models.Post.id == post_id).first()
+    if not post:
+        raise HTTPException(status_code=404, detail="Post not found")
+    return post
+
+
+@router.delete("/{post_id}")
+def delete_post(
+    post_id: int,
+    current_user: models.User = Depends(auth.get_current_user),
+    db: Session = Depends(get_db),
+):
+    post = db.query(models.Post).filter(models.Post.id == post_id).first()
+    if not post:
+        raise HTTPException(status_code=404, detail="Post not found")
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not authorized to delete this post")
+
+    db.delete(post)
+    db.commit()
+    return {"message": "Post deleted successfully"}
